Add config.py to .gitignore and change app config
To prepare to add forms to the app with the Flask-WTF extension,
a config.py file has been added to the project and included in the
.gitignore. Currently, the only thing in config.py is a Config class
that keeps track of the SECRET_KEY for the app.

Finally, Config class is imported to __init.py__ so it can be used
to configure the app.