Restrict page(s) from view unless logged in
Adds functionality of view restricting view functions with the help of
Flask-Login. The @login_required decorator is all that is necessary for
future pages to be view restricted.

Unauthorized users will be redirected to the login page, then back to
the page they were trying to access after successfully logging in.