Fix showing default values in edit profile form
After some research, there are a few ways to go about showing default
values in a FlaskForm. One of the parameters that can be passed to a
FlaskForm is the 'obj' parameter. If the object passed has the same
attribute names as the field names, the form will populate the fields
with the data in those attributes.