Allow for unchanged email in edit profile page
Previously, the form required the email to be different from the default
value of the email field, which would be the email the user currently
has saved. The form checks if no users are currently registered with
that email, raising a validation error. To correct this, must also check
if the given email is different from the default email.